+++ packages/sdk/typescript/src/api/resources/annotations/client/requests/index.ts
@@ -1,1 +1,4 @@
 export type { CreateAnnotationBody } from "./CreateAnnotationBody.js";
+export type { DeleteAnnotationsRequest } from "./DeleteAnnotationsRequest.js";
+export type { GetAnnotationsRequest } from "./GetAnnotationsRequest.js";
+export type { UpdateAnnotationBody } from "./UpdateAnnotationBody.js";